==> code/3_ISG_score_analysis_2.Rmd <==
---
title: "ISG scores"
output: html_notebook
---

### clean environment
```{r}
rm(list=ls())
```

### set directory
```{r setup}
    knitr::opts_knit$set(root.dir = normalizePath("~/Desktop/Code-PHD/IFN_Covid_multiomics/"))  
```

```{r}
library(mygene)
library(org.Hs.eg.db)
library(circlize)
library(ConsensusClusterPlus)
library(ComplexHeatmap)
library(reshape2)
```


```{r}
col_clusters <- c("#FFA500","#A0A0A4","#489048","#bd7f7b","#7b98bd")
```

```{r}
col_conditions <- c("#FFA500","#A0A0A4","#7F7BBD","#C05A12")
```

```{r}
antiviral <- read.delim("data/databases/IFN/antiviral_mechanism_IFN.tsv")
antiviral <- gsub(".* ", "", antiviral$MoleculeName)
inf_alpha_beta <- read.delim("data/databases/IFN/reactome_interferon_alpha_beta.tsv")
inf_alpha_beta <- gsub(".* ", "", inf_alpha_beta$MoleculeName)
inf_gamma <- read.delim("data/databases/IFN/reactome_interferon_gamma.tsv")
inf_gamma <- gsub(".* ", "", inf_gamma$MoleculeName)
inf_lambda <- c("IL29", "IL28A", "IL28B")

inf <- unique(c(antiviral, inf_alpha_beta, inf_gamma, inf_lambda))
```

```{r}
data_txn <- read.csv("processing/COVID_transcriptomics_vsd_data_deseq2.csv", row.names = 1, check.names = FALSE)
data_txn <- data_txn[, sapply(data_txn, var) > 0.2]
data_txn <- data.frame(t(data_txn))
data_txn$Symbol <- mapIds(org.Hs.eg.db, rownames(data_txn), keytype="ENSEMBL", column="SYMBOL")
data_txn <- data_txn[!is.na(data_txn$Symbol),]
data_txn <- data_txn[!duplicated(data_txn$Symbol),]
rownames(data_txn) <- data_txn$Symbol
data_txn <- data_txn[rownames(data_txn) %in% inf,]
data_txn$Symbol <- NULL
data_txn_2 <- data.frame(t(data_txn))
rownames(data_txn_2) <- gsub("\\.", "-", rownames(data_txn_2))
data_txn_2$X <- rownames(data_txn_2)
```
```{r}
clinical <- read.csv("processing/clean_data_clinical.csv")
clinical <- clinical[,c(2, 39, 4, 5)]
names(clinical)[3] <- "Age"

table(clinical$condition)
hiv <- clinical[!clinical$condition %in% c("HC", "Convalescent") & !is.na(clinical$condition),]
names(hiv)[1] <- "X"
names(clinical)[1] <- "X"

hiv <- merge(hiv, data_txn_2, by = "X")
data_covid <- merge(clinical, data_txn_2, by = "X")
```

## clustering based on data
## consensus clustering
```{r}
rcc = ConsensusClusterPlus(t(data_covid[,-c(1:4)]),maxK=10,reps=1000,pItem=0.8,pFeature=1,title="results/clustering/Txn_IFN_all_patients_2",distance="spearman",clusterAlg="hc", plot="pdf")

x <- rcc[[3]]$consensusMatrix
y <- data.frame(rcc[[3]][["consensusClass"]])
#min(x)
#max(x)
```

```{r}
write.csv(x, "consensus_matrix_heatmap_transcriptomics_all_samples.csv")
```

```{r}
col_fun = colorRamp2(c(0.001, 1), c("white", "#FF6F61"))
```

```{r}
data_covid$condition <- factor(data_covid$condition, levels = c("HC", "Conv", "Mild", "Severe"))
```


```{r}
clusters <- read.csv("/home/flomik/Desktop/Code-PHD/IFN_Covid_multiomics/processing/clusters_concensus_clustering.csv")
clusters$cluster <- as.vector(clusters$cluster)
clusters$cluster[clusters$cluster == 1] <- "LIS"
clusters$cluster[clusters$cluster == 2] <- "HIS"
clusters$cluster[clusters$cluster == 3] <- "MIS"
clusters$cluster[clusters$cluster == "HC_1"] <- "HC"
clusters$cluster[clusters$cluster == "Convalescent_1"] <- "Conv"
#clusters$cluster[clusters$X == "COVID19-037"] <- "outlier"
clusters$cluster <- factor(clusters$cluster, levels = c("HC", "Conv", "LIS", "MIS", "HIS"))
```

```{r}
df <- data.frame(X = data_covid$X, condition_bis = y$rcc..3.....consensusClass...)
```

```{r}
df <- merge(df, clusters, by = "X")
```

```{r}
table(df$condition_bis, df$cluster)
```

```{r}
data_covid <- merge(clusters, data_covid, by = "X", all.y = TRUE)
```

```{r}
pdf("results/figures/clustering_Txn_data_all_patients.pdf", height = 6, width = 7)
ht = Heatmap(as.matrix(x))

ha = HeatmapAnnotation(group = data_covid$condition, col = list(group = c("HC" = col_conditions[1], "Conv" = col_conditions[2], "Mild" =col_conditions[3], "Severe" = col_conditions[4])))

ha2 = HeatmapAnnotation(cluster = data_covid$cluster, col = list(cluster = c("HC" = col_conditions[1], "Conv" = col_conditions[2],"LIS" = col_clusters[3], "HIS" = col_clusters[4], "MIS" = col_clusters[5])))

ht = Heatmap(as.matrix(x), name = "Consensus", col = col_fun, top_annotation = ha, bottom_annotation = ha2, show_column_names = FALSE, show_row_names = FALSE, show_column_dend = FALSE, show_row_dend = FALSE)
draw(ht, show_heatmap_legend = TRUE, show_annotation_legend = TRUE)
dev.off()
```

```{r}
x <- table(data_covid$condition, data_covid$cluster)
x <- melt(x)
```

```{r}
library(ggplot2)
# Stacked barplot with multiple groups
ggplot(data=x, aes(x=Var2, y=value, fill=Var1, label = value)) +
  geom_bar(stat="identity")+ scale_fill_manual(values=col_conditions)+theme_classic()+
  geom_text(aes(label=ifelse(value >= 0.10, value,"")), size = 7, position = position_stack(vjust = 0.5), color = "white")

ggsave("results/figures/barplots_condition_clusters.pdf")
```
## PCA / UMAP
```{r}
pca_data <- data_covid[,-c(1, 3:5)]
names(pca_data)[1] <- "condition"
rownames(pca_data) <- data_covid$X
pca_data[is.na(pca_data)] <-  0
pca <- prcomp(pca_data[,-1])
df_out <- as.data.frame(pca$x)
df_out$group <- as.factor(pca_data$condition)
df_out$label <- rownames(pca_data)
x <- project.pca.proportionvariances <- ((pca$sdev^2) / (sum(pca$sdev^2)))*100
a1 <- x[1]
a2 <- x[2]


gg <- data.frame(cluster=factor(df_out$group), x=df_out$PC1, y=df_out$PC2)
centroids <- aggregate(cbind(x,y)~cluster,data=gg,mean)
gg <- merge(gg,centroids,by="cluster",suffixes=c("",".centroid"))

ggplot(gg)+
  geom_point(aes(x=x,y=y,color=cluster, fill = cluster), size=3)+
  stat_ellipse(geom = "polygon", aes(x=x,y=y,color=cluster, fill = cluster), alpha = 0.3) +
  geom_point(data=centroids, aes(x=x, y=y, color=cluster, fill = cluster), size = 5, alpha = 0.9, shape = 19)+ 
  geom_segment(aes(x=x.centroid, y=y.centroid, xend=x, yend=y, color=cluster))+
  theme(legend.title=element_text(size=15),legend.text=element_text(size=12),
        legend.key.size=unit(0.7,"line"),plot.title = element_text(hjust = 0.5),
        axis.title.y=element_text(size=15),axis.title.x=element_text(size=15),
        axis.text.y=element_text(size=18),axis.text.x=element_text(size=18))+ scale_color_manual(values=col_clusters)+ scale_fill_manual(values=col_clusters)+
  xlab(paste0("PC1: ",round(a1,0),"% variance"))+
  ylab(paste0("PC2: ",round(a2,0),"% variance"))+ theme_classic()+
    theme(axis.title.x = element_text(color="black"),
          axis.title.y = element_text(color="black"),
          axis.text = element_text(color="black"),
          axis.text.x =element_text(color="black"),
          panel.border = element_blank(),
          axis.ticks.x = element_blank(),
          strip.text.x = element_text(size = 9,color="black"),
          panel.background = element_rect(fill = "white", colour = "white", size = 0.5, linetype = "solid"),
          plot.title = element_text(hjust = 0.5, size = 20, face = "bold", vjust=2),
          legend.background=element_rect(fill = alpha("white", 0)))
ggsave("results/figures/PCA_new_clusters_ISG.pdf", height = 6, width = 7)
```

```{r}
data_covid <- data_covid[data_covid$cluster != "outlier",]
write.csv(data_covid[,c(1:5)], "processing/clean_data_clinical_new_clusters.csv")
```

```{r}
clinical <- read.csv("processing/clean_data_clinical_new_clusters.csv")
clinical$X.1 <- NULL
clinical$cluster <- factor(clinical$cluster, levels = c("HC", "Conv", "LIS", "MIS", "HIS"))
clinical$condition <- factor(clinical$condition, levels = c("HC", "Conv", "Mild", "Severe"))
```

```{r}
table(data_covid$cluster)
```


```{r}
table(pca_data$condition)
```

```{r}
data_txn$Symbol <- NULL
data_type_I <- data_txn[rownames(data_txn) %in% inf_alpha_beta,]
data_type_I <- data.frame(t(data_type_I))
data_type_I$type_I_score <- apply(data_type_I, 1, sum)
data_type_I$type_I_score <- scale(data_type_I$type_I_score)


data_type_II <- data_txn[rownames(data_txn) %in% inf_gamma,]
data_type_II <- data.frame(t(data_type_II))
data_type_II$type_II_score <- apply(data_type_II, 1, sum)
data_type_II$type_II_score <- scale(data_type_II$type_II_score)

#data_type_III <- data_txn[rownames(data_txn) %in% inf_lambda,] there is none
data_type_antiviral <- data_txn[rownames(data_txn) %in% antiviral,]
data_type_antiviral <- data.frame(t(data_type_antiviral))
data_type_antiviral$type_antiviral_score <- apply(data_type_antiviral, 1, sum)
data_type_antiviral$type_antiviral_score <- scale(data_type_antiviral$type_antiviral_score)

isg_scores <- cbind(data_type_I$type_I_score, data_type_II$type_II_score, data_type_antiviral$type_antiviral_score)
colnames(isg_scores) <- c("type_I", "type_II", "type_antiviral")
rownames(isg_scores)  <- rownames(data_type_I)
write.csv(isg_scores, "results/isg_scores_per_patients.csv")
```

```{r}
isg_scores <- data.frame(isg_scores)
isg_scores$X <- rownames(isg_scores)
isg_scores$X <- gsub("\\.", "-", isg_scores$X)
clinical <- data_covid[,c(1:5)]
clinical <- merge(clinical, isg_scores, by = "X")
clinical$cluster <- as.factor(clinical$cluster)
```

```{r}
library(reshape2)
library(ggpubr)
```

## boxplots ISGs score by
```{r}
pvalues <- data.frame(Var1 = NA, Var2 = NA, value = NA, test = NA)

col_list <- list(cluster = col_clusters,
                 condition = col_conditions)
i_condition <- c(2, 3)
i_isg_scores <- c(6:8)
m = 1

for (i in i_condition) {
  for (j in i_isg_scores) {
    print(names(clinical)[i])
    print(names(clinical)[j])
    table_1 <- clinical[,c(j, i)]
    
    x <- pairwise.wilcox.test(table_1[,1],table_1[,2], data = table_1, p.adjust.method = "fdr")
    x2 <- melt(x[[3]])
    x2$test <- paste0(names(clinical)[i],"_", names(clinical)[j])
    pvalues <- rbind(pvalues, x2)
    list_comp <- as.vector(unique(table_1[,2]))
    list_pair <- data.frame(t(combn(list_comp,2)))
    list_pair$X1 <- as.vector(list_pair$X1)
    list_pair$X2 <- as.vector(list_pair$X2)
    list_pair <- list_pair[!list_pair$X2 == "Conv",]
    comp <- list(c(list_pair$X1[1], list_pair$X2[1]),
                 c(list_pair$X1[2], list_pair$X2[2]),
                 c(list_pair$X1[3], list_pair$X2[3]))
    my_comparisons <- comp
    col <- col_list[[m]]
    a <- ggplot(table_1, aes(y = table_1[,1], x = table_1[,2], fill = table_1[,2])) + 
    geom_boxplot(position=position_dodge(1), size = 0.1, alpha =0.8, outlier.shape = NA)+ 
    labs(y= paste0("IFN ",names(table_1)[1], ""), x = paste0(names(table_1)[2], ""))+ theme_classic()+
    theme(axis.title.x = element_blank(),
          axis.title.y = element_text(color="black"),
          axis.text = element_text(color="black"),
          axis.text.x =element_text(color="black"),
          panel.border = element_blank(),
          axis.ticks.x = element_blank(),
          strip.text.x = element_text(size = 9,color="black"),
          legend.position="none",
          panel.background = element_rect(fill = "white", colour = "white", size = 0.5, linetype = "solid"),
          plot.title = element_text(hjust = 0.5, size = 20, face = "bold", vjust=2),
          legend.background=element_rect(fill = alpha("white", 0))) + 
    geom_jitter(shape=16, size=1, color="black", position=position_jitter(0.05))+ 
  stat_compare_means(comparisons = my_comparisons)+ scale_fill_manual(values=col)
    ggsave(paste0("results/figures/boxplots/", names(clinical)[i],"_", names(clinical)[j],"_boxplots.pdf"), width = 3.5, height = 2.5)
  }
  m = m + 1 
}


write.csv(pvalues, "processing/mann_withney_ISG_scores_all_together.csv")
pvalues <- pvalues[pvalues$Var1 != "Conv" & pvalues$Var2 != "Conv",]
pvalues <- pvalues[!is.na(pvalues$value),]
write.csv(pvalues, "processing/mann_withney_ISG_scores_all_together_filt.csv")
```
