==> code/7_clinical_data_COVID.Rmd <==
---
title: "Clinical data COVID19"
output: html_notebook
---


### clean environment
```{r}
rm(list=ls())
```

### set directory
```{r setup}
    knitr::opts_knit$set(root.dir = normalizePath("~/Desktop/Code-PHD/IFN_Covid_multiomics/"))  
```

```{r}
library(dplyr)
library(xlsx)
```

## clean clinical data
```{r}
clinical <- read.csv("processing/clean_data_clinical.csv")
```

```{r}
clinical$Gender <- as.vector(clinical$Gender)
clinical$Gender[clinical$Gender == "male"] <- "Male"

clinical$Ethinicity <- as.vector(clinical$Ethinicity)
clinical$Ethinicity[is.na(clinical$Ethinicity)] <- "Other"
clinical$Ethinicity[clinical$Ethinicity == "Asian"] <- "Other"

clinical$O2.demand.inclusion..litre.min. <- as.vector(clinical$O2.demand.inclusion..litre.min.)
clinical$O2.demand.inclusion..litre.min.[is.na(clinical$O2.demand.inclusion..litre.min.)] <- 0
clinical$O2.demand.inclusion..litre.min.[clinical$O2.demand.inclusion..litre.min. == "80% m HFNO 30L"] <- 0.8*30
clinical$O2.demand.inclusion..litre.min. <- as.numeric(clinical$O2.demand.inclusion..litre.min.)


clinical$No..Of.days.between.symptom.onset.to.sample.collection[is.na(clinical$No..Of.days.between.symptom.onset.to.sample.collection)] <- 0

clinical$Comorbidities..Diabetes..CVD..COPD..Others...Gr.1.22 <- as.vector(clinical$Comorbidities..Diabetes..CVD..COPD..Others...Gr.1.22)

clinical$Comorbidities_2 <- NA
clinical$Comorbidities_2[!is.na(clinical$Comorbidities..Diabetes..CVD..COPD..Others...Gr.1.22)] <- c("Other", NA, NA, "CVD+Other", NA, NA, "COPD", "Diabete+CVD", NA, NA, "CVD", "CVD+Other", "Other", "CVD+Other", "Other", "CVD", "COPD+Other", NA, "Diabete+CVD+Other", "Other", "Diabete+CVD", "CVD+COPD", "CVD+COPD", "CVD+Other", NA)
```

```{r}
clinical$Comorbidities_3 <- NA
clinical$Comorbidities_3[!is.na(clinical$Comorbidities_2)] <- "yes"


clinical$Comorbidities_3 <- as.factor(clinical$Comorbidities_3)
clinical$Gender <- as.factor(clinical$Gender)
clinical$Ethnicity <- as.factor(clinical$Ethinicity)

#clinical$Comorbidities_2 <- NULL
clinical$Comorbidities_1 <- NULL
```

```{r}
names(data)
```
```{r}
clinical$KI_ID[is.na(clinical$Date.when.onset.of.symptoms...for.Gr.1.2.) & !grepl("HC", clinical$KI_ID)]
```
```{r}
clinical$KI_ID[is.na(clinical$COVID.Diagnosis.Date...for.Gr.1.2.) & !grepl("HC", clinical$KI_ID)]
```
```{r}
clinical$KI_ID[is.na(clinical$Hospitalization.date...for.Gr.1.2.) & !grepl("HC", clinical$KI_ID)]
```
```{r}
clinical$KI_ID[is.na(clinical$No..Of.days.between.symptom.onset.to.sample.collection) & !grepl("HC", clinical$KI_ID)]
```

```{r}
data <- clinical
data <- data[!data$condition %in% c("HC", "Conv"),]
data <- data[,c(2, 4, 5, 12:20, 39, 40)]
names(data)
data$Sample_After.hospiytalization <- NULL
data <- data[!is.na(data$Date.when.onset.of.symptoms...for.Gr.1.2.),]
data$date_symptoms <- as.Date(data$Date.when.onset.of.symptoms...for.Gr.1.2.)
data$date_diagnosis <- as.Date(data$COVID.Diagnosis.Date...for.Gr.1.2.)
data$date_hospital <- as.Date(data$Hospitalization.date...for.Gr.1.2.)
data$date_sample_collection <- data$date_symptoms + data$No..Of.days.between.symptom.onset.to.sample.collection
data$Age <- ifelse(data$Age..as.of.2020.> 60, ">60", "<60")
data <- data[,c(1, 3, 4, 12:18)]
names(data)
data <- data[order(data$condition, decreasing = TRUE),]
data$KI_ID <- as.vector(data$KI_ID)
data$KI_ID <- factor(as.vector(data$KI_ID), levels = as.vector(data$KI_ID))
data$Comorbidities_2 <- NULL
```

```{r}
library(reshape2)
library(ggplot2)

```

```{r}
data_heatmap <- data
data_heatmap$BMI_cat <- NA
data_heatmap$BMI_cat[data_heatmap$BMI < 18.5] <- "underweight"
data_heatmap$BMI_cat[data_heatmap$BMI >= 18.5 & data_heatmap$BMI < 25] <- "normal"
data_heatmap$BMI_cat[data_heatmap$BMI >= 25 & data_heatmap$BMI < 30] <- "overweight"
data_heatmap$BMI_cat[data_heatmap$BMI >= 30 & data_heatmap$BMI < 35] <- "obese"
data_heatmap$BMI_cat[data_heatmap$BMI >= 35] <- "extremely obese"

data_heatmap$BMI_cat <- factor(data_heatmap$BMI_cat, levels = c("underweight", "normal", "overweight", "obese", "extremely obese"))

data_heatmap$BMI <- NULL
data_2 <- melt(data_heatmap)
```

```{r}
data_2$KI_ID_2 <- as.vector(data_2$KI_ID)
data_2$KI_ID_2 <- gsub("COVID19.", "CV", data_2$KI_ID_2)
data_2 <- data_2[order(data_2$condition, decreasing = TRUE),]
data_2$KI_ID_2 <- factor(as.vector(data_2$KI_ID_2), levels = as.vector(unique(data_2$KI_ID_2)))
```

```{r}
library("extrafont")

ggplot(data = data_2, aes(group = KI_ID, x = value, y = KI_ID_2, color = condition, fill = variable))+
  geom_line(size = 2, alpha = 0.3) + 
  geom_point(size = 6, shape = 21)+
  theme_classic() + xlim(as.Date("2020-04-15"),as.Date("2020-07-6"))+scale_color_manual(values = c("#7F7BBD","#C05A12"))+scale_fill_manual(values = c("#aab597", "#768858", "#556b2f", "#33401c"))+theme(text=element_text(family="sans"))
  
ggsave("results/figures/time_line_covid_patients_sans.pdf", height = 8, width = 9)
```
```{r}
data_heatmap$value <- c(1:36)
data_heatmap$value_2 <- c(36:1)
```

```{r}
col_conditions <- c("#FFA500","#A0A0A4","#7F7BBD","#C05A12")
```

```{r}
rownames(data_heatmap) <- data_heatmap$KI_ID
```

```{r}
data_heatmap <- data_heatmap[order(match(data_heatmap[,1],rev(data_2[,1]))),]
```
 	 	
```{r}
library(ComplexHeatmap)
pdf("results/figures/Heatmat_patients_characteristics.pdf", width = 10, height = 6)
ht = Heatmap(as.matrix(data_heatmap[,c(10:11)]),
             column_order = colnames(data_heatmap[,c(10:11)]),
             row_order = rownames(data_heatmap[,c(10:11)]),
             show_row_names = TRUE,
             show_column_names = TRUE,
             row_title_rot = 0,
             border = FALSE,
             column_gap = unit(8, "mm"), 
             right_annotation = rowAnnotation(condition = data_heatmap$condition,
                                             Gender = data_heatmap$Gender,
                                              Age = data_heatmap$Age,
                                             BMI = data_heatmap$BMI_cat,
                                              col = list(condition = c(
                                                                       "Mild" = col_conditions[3],
                                                                       "Severe" = col_conditions[4]),
                                                         Gender = c("Male" = "blue", "Female" = "red"),
                                                         Age = c("<60" = "#e8a97d", ">60" = "#e87dbd"),
                                                         BMI = c("overweight" = "#006464", "normal" = "#009090", "obese" = "#003939", "underweight" = "#99d2d2", "extremely obese"  = "#000000")))
             )
ht
dev.off()
```
## barplot patients
```{r}
p<-ggplot(data=df, aes(x=dose, y=len)) +
  geom_bar(stat="identity")
p
```

```{r}
clinical <- clinical[, c(2, 4, 6:9, 14, 20, 21)]
clinical <- clinical[!is.na(clinical$V2),]
names(clinical) <- c("KI_ID", "condition", "Age", "Gender", "Ethnicity", "O2", "BMI", "Days_symptoms", "Comorbidities_1", "Comorbidities_2")
```



```{r}
# Function returning compact column classes
col_classes <- function(df) {
    t(as.data.frame(lapply(df, function(x) paste(class(x), collapse = ','))))
}

# Return example data's column classes

```

```{r}
data <- clinical
data$KI_ID <- NULL
data_factor <- data[, c(which(col_classes(data) == "factor"))]
data_numeric <- data[, c(1, which(col_classes(data) != "factor"))]
names(data_numeric)[1] <- "cluster"
```


```{r}
medIQR <- function(x, na.rm = TRUE, ...) {
  qq <- stats::quantile(x, c(0.50, 0.25, 0.75), na.rm = na.rm, ...)
  structure(qq, 
            names = c("Median", "Q1", "Q3"),
            class = "medIQR") # how about EstInterval? smry?
}

medIQR.formula <- function(formula,
                           data = NULL, ...,
                           subset, na.action = NULL,
                           overall = FALSE,
                           overall.label = "Overall") {
  if (missing(formula) || (length(formula) != 3)) 
    stop("'formula' missing or incorrect")
  m <- match.call(expand.dots = FALSE)
  if (is.matrix(eval(m$data, parent.frame()))) 
    m$data <- as.data.frame(data)
  m$... <- m$overall <- m$overall.label <- NULL
  m$na.action <- na.action
  require(stats, quietly = TRUE)
  m[[1L]] <- as.name("model.frame")
  mf <- eval(m, parent.frame())
  response <- attr(attr(mf, "terms"), "response")
  s <- split(mf[[response]], mf[-response])
  
  if(overall) {
    if(any(names(s) == overall.label))
    stop(sQuote(overall.label), " is an existing level")
    s <- c(Overall = list(unlist(s)), s)
  }
  structure(sapply(s, FUN = medIQR, ...),
            class = "medIQR")
}
```

```{r}
length(unique(data$condition))
```

```{r}
names(data_factor)
```
```{r}
names(data_numeric)
```
```{r}
data_numeric$Days_symptoms[data_numeric$Days_symptoms == 0] <- NA
data_numeric <- data_numeric[,c(1, 2, 4)]
```

### Numeric test normality

```{r}
library(nortest)
#library(MVN)
```
data_numeric
## test normality
```{r}
j = 1
list_norm <- list()
for (i in 3:ncol(data_numeric)){
  c1 <- data_numeric[data_numeric$cluster == unique(data$condition)[1],i]
  c2 <- data_numeric[data_numeric$cluster == unique(data$condition)[2],i]
  c3 <- data_numeric[data_numeric$cluster == unique(data$condition)[3],i]
  c4 <- data_numeric[data_numeric$cluster == unique(data$condition)[4],i]
  c5 <- data_numeric[data_numeric$cluster == unique(data$condition)[5],i]
  c6 <- data_numeric[data_numeric$cluster == unique(data$condition)[6],i]
  test_c1 <- as.numeric(shapiro.test(c1)$p.value)
  test_c2 <- as.numeric(shapiro.test(c2)$p.value)
  test_c3 <- as.numeric(shapiro.test(c3)$p.value)
  test_c4 <- as.numeric(shapiro.test(c4)$p.value)
  test_c5 <- as.numeric(shapiro.test(c5)$p.value)
  test_c6 <- as.numeric(shapiro.test(c6)$p.value)
  #print(c(names(data_numeric)[i], test_c1, test_c2, test_c3, test_c4, test_c5))
  if(test_c1 < 0.05 & test_c2 < 0.05 & test_c3 < 0.05 & test_c4 < 0.05 & test_c5 < 0.05 & test_c6 < 0.05){
    print(paste0(names(data_numeric)[i], "is following normal distribution"))
    list_norm[[j]] <- names(data_numeric)[i]
    j = j + 1
  }
}
```
## normally distributed
```{r}
data_norm <- select(data_numeric, cluster, unlist(list_norm))
#data_numeric <- data_numeric[,! colnames(data_numeric) %in% unlist(list_norm)]
```

## global test
```{r}
table_numeric_2 <- data.frame(parameter = rep(names(data_norm)[2:ncol(data_norm)], each = 2), cluster = rep(c(1, 2), length(names(data_norm)[2:ncol(data_norm)])), Mean_SD = NA, pval_anova = NA)
j = 1

for (i in 2:ncol(data_norm)){
  c1 <- data_norm[data_norm$cluster == 1,i]
  c2 <- data_norm[data_norm$cluster == 2,i]
  #c3 <- data_norm[data_norm$cluster == 3,i]
  #c4 <- data_norm[data_norm$cluster == 4,i]
  #c5 <- data_norm[data_norm$cluster == 5,i]
  x <- paste0("Mean =",round(mean(c1, na.rm=TRUE), 2), ", SD =", round(sd(c1, na.rm = TRUE), 2))
  y <- paste0("Mean =",round(mean(c2, na.rm=TRUE), 2), ", SD =", round(sd(c2, na.rm = TRUE), 2))
  #z <- paste0("Mean =",round(mean(c3, na.rm=TRUE), 2), ", SD =", round(sd(c3, na.rm = TRUE), 2))
  #a <- paste0("Mean =",round(mean(c4, na.rm=TRUE), 2), ", SD =", round(sd(c4, na.rm = TRUE), 2))
  #b <- paste0("Mean =",round(mean(c5, na.rm=TRUE), 2), ", SD =", round(sd(c5, na.rm = TRUE), 2))
  table_numeric_2[c(j, j+1), 3] <- c(x, y)
  test <- t.test(c1, c2)
  table_numeric_2[j, 4] <- test$p.value
  j = j + 2
}


table_numeric_2$X <- paste0("C", table_numeric_2$cluster, "_", table_numeric_2$parameter)

table_numeric_2$X <- NULL

write.xlsx(table_numeric_2, "results/clinical/COCOMO_3_layers_normal_distrib_numeric_values_ANOVA_3.xlsx")
```

```{r}
data_numeric$COCOMO_ID <- NULL
```

```{r}
medIQR <- function(x, na.rm = TRUE, ...) {
  qq <- stats::quantile(x, c(0.50, 0.25, 0.75), na.rm = na.rm, ...)
  structure(qq, 
            names = c("Median", "Q1", "Q3"),
            class = "medIQR") # how about EstInterval? smry?
}

medIQR.formula <- function(formula,
                           data = NULL, ...,
                           subset, na.action = NULL,
                           overall = FALSE,
                           overall.label = "Overall") {
  if (missing(formula) || (length(formula) != 3)) 
    stop("'formula' missing or incorrect")
  m <- match.call(expand.dots = FALSE)
  if (is.matrix(eval(m$data, parent.frame()))) 
    m$data <- as.data.frame(data)
  m$... <- m$overall <- m$overall.label <- NULL
  m$na.action <- na.action
  require(stats, quietly = TRUE)
  m[[1L]] <- as.name("model.frame")
  mf <- eval(m, parent.frame())
  response <- attr(attr(mf, "terms"), "response")
  s <- split(mf[[response]], mf[-response])
  
  if(overall) {
    if(any(names(s) == overall.label))
    stop(sQuote(overall.label), " is an existing level")
    s <- c(Overall = list(unlist(s)), s)
  }
  structure(sapply(s, FUN = medIQR, ...),
            class = "medIQR")
}
```

```{r}
table_numeric <- data.frame(parameter = rep(names(data_numeric)[2:ncol(data_numeric)], each = 2), cluster = rep(c(1, 2), length(names(data_numeric)[2:ncol(data_numeric)])), Mean_SD = NA, pval_anova = NA)
j = 1

for (i in 2:ncol(data_numeric)){
  c1 <- data_numeric[data_numeric$cluster == 1,i]
  c2 <- data_numeric[data_numeric$cluster == 2,i]
  #c3 <- data_numeric[data_numeric$cluster == 3,i]
  #c4 <- data_numeric[data_numeric$cluster == 4,i]
  #c5 <- data_numeric[data_numeric$cluster == 5,i]
  x <- paste0(round(medIQR(c1)[[1]], 2), "(", round(medIQR(c1)[[2]], 2), "-", round(medIQR(c1)[[3]], 2), ")")
  y <- paste0(round(medIQR(c2)[[1]], 2), "(", round(medIQR(c2)[[2]], 2), "-", round(medIQR(c2)[[3]], 2), ")")
  #z <- paste0("Mean =",round(mean(c3, na.rm=TRUE), 2), ", SD =", round(sd(c3, na.rm = TRUE), 2))
  #a <- paste0("Mean =",round(mean(c4, na.rm=TRUE), 2), ", SD =", round(sd(c4, na.rm = TRUE), 2))
  #b <- paste0("Mean =",round(mean(c5, na.rm=TRUE), 2), ", SD =", round(sd(c5, na.rm = TRUE), 2))
  table_numeric[c(j, j+1), 3] <- c(x, y)
  test <- wilcox.test(c1, c2)
  print(test)
  table_numeric[j, 4] <- test$p.value
  j = j + 2
}




table_numeric$X <- paste0("C", table_numeric$cluster, "_", table_numeric$parameter)

#table_numeric <-  merge(table_numeric, reg, by = "X")

table_numeric$X <- NULL

write.xlsx(table_numeric, "results/clinical/COCOMO_3_layers_non_norm_numeric_values_kruskal_3.xlsx")
```

```{r}
table(data$cluster)
```

```{r}
data_factor$cluster <- as.factor(as.vector(data_factor$cluster))
```

```{r}
table_factor <- data.frame(Var1 = NA, Var2 = NA, Level = NA, Freq= NA, Freq2 = NA, Parameter = NA, test = NA, pval = NA)
j = 1

for (i in 2:ncol(data_factor)) {
  print(names(data_factor)[i])
  x <- data.frame(t(table(data_factor$cluster, data_factor[,i])))
  x$Level <- x$Var1
  x$Var1 <- paste0(names(data_factor)[i], "_", x$Var1, "_", x$Var2)
  x$Freq2 <- ""
  x$Freq2[x$Var2 == 1] <- paste0(x$Freq[x$Var2 == 1], "(", round(x$Freq[x$Var2 == 1]/sum(x$Freq[x$Var2 == 1])*100, 2), "%)")
  x$Freq2[x$Var2 == 2] <- paste0(x$Freq[x$Var2 == 2], "(", round(x$Freq[x$Var2 == 2]/sum(x$Freq[x$Var2 == 2])*100, 2), "%)")
  #x$Freq2[x$Var2 == 3] <- paste0(x$Freq[x$Var2 == 3], "(", round(x$Freq[x$Var2 == 3]/sum(x$Freq[x$Var2 == 3])*100, 2), "%)")
  #x$Freq2[x$Var2 == 4] <- paste0(x$Freq[x$Var2 == 4], "(", round(x$Freq[x$Var2 == 4]/sum(x$Freq[x$Var2 == 4])*100, 2), "%)")
  #x$Freq2[x$Var2 == 5] <- paste0(x$Freq[x$Var2 == 5], "(", round(x$Freq[x$Var2 == 5]/sum(x$Freq[x$Var2 == 5])*100, 2), "%)")
  x$Parameter <- names(data_factor)[i]
  m1 <- rbind(table(data_factor$cluster, data_factor[,i]))
  f <- chisq.test(m1)
  print(f$expected)
  if(sum(as.vector(f$expected > 5), na.rm = TRUE) == length(as.vector(f$expected))){
    x[1, 7] <- "Chi-Square Test"
    print(x[1, 7])
    x[1, 8] <- f$p.value
    }else{
      x[1, 7] <- "Fisher's Exact Test"
      print(x[1, 7])
      x[1, 8] <- fisher.test(m1, simulate.p.value=TRUE)$p.value
    }
  names(x)[7:8] <- c("test", "pval")
  table_factor <- rbind(x,table_factor)
}

table_factor <- select(table_factor, Var2, Parameter, Level, Freq2, test, pval)
names(table_factor)[1] <- "cluster"
write.xlsx(table_factor, "results/clinical/X2_results_results_cocomo_3_groups_3.xlsx")

```
```{r}
x <- table_numeric$parameter[table_numeric$pval_anova < 0.05 & !is.na(table_numeric$pval_anova)]
x
```

```{r}
table_numeric_2$parameter[table_numeric_2$pval_anova < 0.05 & !is.na(table_numeric_2$pval_anova)]
```


```{r}
y <- table_factor$Parameter[table_factor$pval < 0.05 & !is.na(table_factor$pval)]
y
```

```{r}
write.csv(y, "processing/factor_differing_groups_3.csv")
write.csv(x, "processing/numeric_differing_groups_3.csv")
```
