==> code/8_cell_proportions_FACS_clusters.Rmd <==
---
title: "Figures transcriptomics"
output: html_notebook
---


### clean environment
```{r}
rm(list=ls())
```

### set directory
```{r setup}
    knitr::opts_knit$set(root.dir = normalizePath("~/Desktop/Code-PHD/IFN_Covid_multiomics/"))  
```

### load packages
```{r}
library(DESeq2)
library(dplyr)
library(SNFtool)
library(xlsx)
library(ggplot2)
library(corrr)
library(mygene)
library(edgeR)
library(xlsx)
library(reshape2)
```

```{r}
col <- c("red","blue", "yellow", "pink")
```

```{r}
data <- read.csv("data/MNP_COVID19/MNP Data compiled.csv")
data <- data[,c(1, 2, 5:14)]
colnames(data)
rownames(data) <- data$Sample.
```

```{r}
list_comp <- unique(data$COVID.status)
data$X <- gsub("\\_D.*", "",data$Sample.)
data$Sample. <- NULL
data$X <- gsub("\\_C.*", "",data$X)
data$X <- gsub("\\_B.*", "",data$X)
data$X <- gsub("\\_", "-",data$X)
data$X <- gsub("CoV19", "COVID19",data$X)

data$X[!grepl("HC", data$X)] <- gsub("\\-", "-0",data$X[!grepl("HC", data$X)])
```

```{r}
clinical <- read.csv("processing/clean_data_clinical_new_clusters.csv")
clinical$X
data_2 <- merge(clinical, data, by = "X")
```


```{r}
write.csv(data_2, "processing/data_MNP_FACS_processed.csv")
```

```{r}
comp <- list(c("LIS", "HIS"), c("LIS", "MIS"), c("HIS", "MIS"), c("HC", "LIS"), c("HC", "HIS"), c("HC", "MIS"), c("HC", "Conv"))
```

```{r}
fractions <- data_2
```

```{r}
names(data)
```

```{r}
data <- data_2
```

```{r}
table_fraction <- data.frame(Var1 = NA, Var2 = NA, value = NA, Cell_line = NA)

for (i in 8:ncol(data)) {
  data_1 <- data[,c(3, i)]
  x <- pairwise.wilcox.test(data_1[,2], data_1$cluster, p.adj = "fdr")
  x2 <- melt(x[[3]])
  x2$Cell_line <- colnames(data_1)[2]
  table_fraction_2 <- x2
  table_fraction <- rbind(table_fraction_2, table_fraction)
}

table_fraction <- table_fraction[complete.cases(table_fraction),]
table_fraction$Comp <- paste0(table_fraction$Var1, "_", table_fraction$Var2)
table_fraction$Var1 <- NULL
table_fraction$Var2 <- NULL
table_fraction_2 <- dcast(table_fraction, Comp ~ Cell_line)
write.csv(table_fraction_2, "results/MNP_differences_proportions_clusters.csv")
rownames(table_fraction_2) <- table_fraction_2$Comp
table_fraction_2$Comp <- NULL
table_fraction_2[table_fraction_2 > 0.1] <- NA
write.csv(table_fraction_2, "results/MNP_differences_proportions_clusters_filt.csv")
```

```{r}
table_fraction <- data.frame(Var1 = NA, Var2 = NA, value = NA, Cell_line = NA)

for (i in 8:ncol(data)) {
  data_1 <- data[,c(4, i)]
  x <- pairwise.wilcox.test(data_1[,2], data_1$condition, p.adj = "fdr")
  x2 <- melt(x[[3]])
  x2$Cell_line <- colnames(data_1)[2]
  table_fraction_2 <- x2
  table_fraction <- rbind(table_fraction_2, table_fraction)
}

table_fraction <- table_fraction[complete.cases(table_fraction),]
table_fraction$Comp <- paste0(table_fraction$Var1, "_", table_fraction$Var2)
table_fraction$Var1 <- NULL
table_fraction$Var2 <- NULL
table_fraction_2 <- dcast(table_fraction, Comp ~ Cell_line)
write.csv(table_fraction_2, "results/MNP_differences_proportions_conditions.csv")
rownames(table_fraction_2) <- table_fraction_2$Comp
table_fraction_2$Comp <- NULL
table_fraction_2[table_fraction_2 > 0.1] <- NA
write.csv(table_fraction_2, "results/MNP_differences_proportions_conditions_filt.csv")
```


```{r}
colnames(data)
```

```{r}
table_1 <- data[,c(8, 3)]
```

```{r}
col_clusters <- c("#FFA500","#A0A0A4","#a07bbd","#7b98bd", "#bd7f7b")
```

## make dotplot
```{r}
table_1$cluster <- as.vector(table_1$cluster)
#table_1$cluster[table_1$cluster == "HC_1"] <- "HC"
table_1$cluster[table_1$cluster == "Convalescent_1"] <- "Conv"
table_1$cluster <- factor(table_1$cluster, levels = c("HC", "Conv", "LIS", "MIS", "HIS"))
list_comp <- as.vector(unique(table_1[,2]))
list_pair <- data.frame(t(combn(list_comp,2)))
list_pair$comp <- paste0(list_pair$X1, "-", list_pair$X2)
comp <- list_pair$comp
comp <- comp[!grepl("Convalescent", comp) & !grepl("HC_1-Convalescent_1", comp)]
my_comparisons <- comp
a <- ggplot(table_1, aes(y = table_1[,1], x = table_1[,2], fill = table_1[,2]))+
  geom_dotplot(binaxis='y', stackdir='center', alpha =0.8, dotsize = 1.5) + 
    labs(y= paste0("",names(table_1)[1], "(%)"), x = paste0(names(table_1)[2], ""))+ theme_classic()+
    theme(axis.title.x = element_blank(),
          axis.title.y = element_text(size=16, face = "bold",color="black"),
          axis.text = element_text(color="black",size=16, face = "bold"),
          axis.text.x =element_text(color="black",size=16, face = "bold"),
          panel.border = element_blank(),
          axis.ticks.x = element_blank(),
          strip.text.x = element_text(size = 9,color="black"),
          legend.position="none",
          panel.background = element_rect(fill = "white", colour = "white", size = 0.5, linetype = "solid"),
          plot.title = element_text(hjust = 0.5, size = 20, face = "bold", vjust=2),
          legend.background=element_rect(fill = alpha("white", 0)))+scale_fill_manual(values = c(col_clusters))
a
ggsave(paste0("results/figures/dotplots_Neutrophils_MNP.pdf"), width = 4, height = 3)
```
```{r}
fractions <- data[,-c(1:3,5:7)]
fractions_2 <- aggregate(. ~ condition, data=fractions, FUN=mean)
```


```{r}
phylum_colors <- c("#5F7FC7","#DA5724", "#508578", "#CD9BCD",
   "#AD6F3B", "#673770","#D14285", "#652926", "#C84248", 
  "#8569D5", "orange", "#CBD588", "#edf2ef", "#5E738F","#D1A33D", "#8A7C64")
```

```{r}
fractions_2[1,-1] <- fractions_2[1,-1]/sum(fractions_2[1,-1])
fractions_2[2,-1] <- fractions_2[2,-1]/sum(fractions_2[2,-1])
fractions_2[3,-1] <- fractions_2[3,-1]/sum(fractions_2[3,-1])

fractions_2[4,-1] <- fractions_2[4,-1]/sum(fractions_2[4,-1])
fractions_2[5,-1] <- fractions_2[5,-1]/sum(fractions_2[5,-1])
```

```{r}
library(reshape2)
fractions_3 <- melt(fractions_2)
```
```{r}
fractions_3 <- fractions_3[!is.na(fractions_3$condition),]
```
```{r}
fractions_3$condition <- factor(fractions_3$condition, levels = c("HC", "Conv", "Mild", "Severe"))
```
## mean barplot by cluster
```{r}
m1 <- ggplot(fractions_3, aes(fill=variable, y=value, x=condition)) + 
    geom_bar(position="fill", stat="identity", alpha= 0.6)+ylab("")+ theme_void()+ 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text = element_blank(),
          panel.border = element_blank(),
          axis.ticks.x = element_blank(),
          strip.text.x = element_text(size = 7,color="black"),
          legend.position="bottom",
          axis.text.x = element_text(size = 7,color="black")
          )+ theme(legend.position = "right")+scale_fill_manual(values = c(phylum_colors, "red", "blue")) + geom_text(aes(label=ifelse(value >= 0.10, paste0(round(value, 2), ""),"")), position=position_stack(vjust=0.5), colour="black")

m1


ggsave("results/figures/MNP_proportions_condition.pdf", height = 3, width = 6)
```

```{r}
fractions <- data[,-c(1,2, 4:7)]
fractions_2 <- aggregate(. ~ cluster, data=fractions, FUN=mean)
```


```{r}
phylum_colors <- c("#5F7FC7","#DA5724", "#508578", "#CD9BCD",
   "#AD6F3B", "#673770","#D14285", "#652926", "#C84248", 
  "#8569D5", "orange", "#CBD588", "#edf2ef", "#5E738F","#D1A33D", "#8A7C64")
```

```{r}
fractions_2[1,-1] <- fractions_2[1,-1]/sum(fractions_2[1,-1])
fractions_2[2,-1] <- fractions_2[2,-1]/sum(fractions_2[2,-1])
fractions_2[3,-1] <- fractions_2[3,-1]/sum(fractions_2[3,-1])

fractions_2[4,-1] <- fractions_2[4,-1]/sum(fractions_2[4,-1])
fractions_2[5,-1] <- fractions_2[5,-1]/sum(fractions_2[5,-1])
```

```{r}
library(reshape2)
fractions_3 <- melt(fractions_2)
```

```{r}
fractions_3$condition <- factor(fractions_3$cluster, levels = c("HC", "Conv", "LIS", "MIS", "HIS"))
```

## mean barplot by cluster
```{r}
m1 <- ggplot(fractions_3, aes(fill=variable, y=value, x=condition)) + 
    geom_bar(position="fill", stat="identity", alpha= 0.6)+ylab("")+ theme_void()+ 
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text = element_blank(),
          panel.border = element_blank(),
          axis.ticks.x = element_blank(),
          strip.text.x = element_text(size = 7,color="black"),
          legend.position="bottom",
          axis.text.x = element_text(size = 7,color="black")
          )+ theme(legend.position = "right")+scale_fill_manual(values = c(phylum_colors, "red", "blue")) + geom_text(aes(label=ifelse(value >= 0.10, paste0(round(value, 2), ""),"")), position=position_stack(vjust=0.5), colour="black")

m1


ggsave("results/figures/MNP_proportions_clusters.pdf", height = 3, width = 6)
```

## bubble plot significant
```{r}
cell <- table_fraction$cell_type[c(1, 2, 5, 6, 8,10)]

fractions_4 <- fractions_3[fractions_3$variable %in% cell,]
```

```{r}
fractions_3$variable
```

```{r}
fractions_4 <- fractions_4[fractions_4$value > 0,]

m1 <- ggplot(fractions_4, aes(y=variable, x=condition, size = value, col = value))+geom_point()+
  scale_size_continuous(range = c(2,6))+
   scale_colour_gradient(low = "#6074a1", high = "red", na.value = NA)+theme_classic()
m1
ggsave("results/figures/proportions_cell_types_bulk_clusters.pdf", height = 4, width = 6)
```
